chore: useButton hook to handle buttons
