+++ src/types/types.ts
@@ -13,4 +13,12 @@ export interface TimeContextInterface {
   setTimerActive: (arg0: boolean) => void;
   toggleTimer: () => void;
   resetTimer: () => void;
+  setValidTime: (arg0: boolean) => void;
+  setSeconds: (arg0: number) => void;
+}
+
+export interface ButtonState {
+  presetTime: (e: React.MouseEvent<HTMLButtonElement, MouseEvent>) => void;
+  toggleTimer: () => void;
+  resetTimer: () => void;
 }
